refactor: app/game.tsx aufteilen in Phase-Komponenten (Issue #202)
- MemorizePhase, DrawPhase, ResultPhase nach components/game/ extrahiert
- game.tsx von 1270 auf 438 LOC reduziert (Orchestrator-Pattern)
- game.shared.ts mit TypeScript-Props-Typen für alle drei Phasen
- 14 neue Tests (min. 4 je Komponente), bestehende Tests bleiben grün

Closes #202
