+++ components/game/game.shared.ts
@@ -1,0 +1,62 @@
+import type { DrawingPath } from '@components/DrawingCanvas';
+import type { LevelImage } from '../../types';
+
+export interface MemorizePhaseProps {
+  timeRemaining: number;
+  currentImage: LevelImage | null;
+  levelNumber: number;
+  currentLang: string;
+  memorizeImageSize: number;
+  imagePlaceholderMinSize: number;
+  revealStep: number;
+}
+
+export interface DrawPhaseProps {
+  currentImage: LevelImage | null;
+  currentLang: string;
+  hasUsedHint: boolean;
+  onUseHint: () => void;
+  onShowHintModal: () => void;
+  drawing: {
+    color: string;
+    strokeWidth: number;
+    tool: 'brush' | 'fill';
+    paths: DrawingPath[];
+    setPaths: (paths: DrawingPath[]) => void;
+    setColor: (color: string) => void;
+    setTool: (tool: 'brush' | 'fill') => void;
+    setStrokeWidth: (width: number) => void;
+    undo: () => void;
+  };
+  layout: {
+    canvasMaxHeight: number;
+    canvasMinHeight: number;
+    canvasMarginVertical: number;
+    toolbarMarginVertical: number;
+    toolbarButtonMinHeight: number;
+    toolbarButtonPaddingVertical: number;
+    buttonMinHeight: number;
+    buttonPaddingVertical: number;
+    isSmall: boolean;
+  };
+  onDone: () => void;
+}
+
+export interface ResultPhaseProps {
+  currentImage: LevelImage | null;
+  currentLang: string;
+  levelNumber: number;
+  userRating: number;
+  savedToGallery: boolean;
+  isReplaying: boolean;
+  replayPaths: DrawingPath[];
+  drawingPaths: DrawingPath[];
+  screenWidth: number;
+  isSmall: boolean;
+  onRatingSubmit: (rating: number) => void;
+  onSaveToGallery: () => void;
+  onStartReplay: () => void;
+  onStopReplay: () => void;
+  onNextLevel: () => void;
+  onRestartFromLevel1: () => void;
+}
